Upload Stage 2 v1 (index_v3)
This is the upload of the files from the end of Stage 2 (Colours and Fonts). In all it is the third version of my website. I will do another upload and pull request with some stakeholder feedback for this stage. That will make the fourth version of my website, named in Projectspace as index_v4.html and style_v4.css.