Upload Stage 2 v2 (index_v4)
This is the upload of the changes after getting stakeholder feedback at the end of Stage 2. It is Stage 2 v2; the second set of files for Colours and Fonts. It is the fourth total version of the website (v4), and the Projectspace equivalent is index_v4.html, style_v4.css, and script_v1.js