Upload Stage 3 v1 (index_v5)
This is the upload of the files from the fist part of Stage 3 (Images and Graphics). In all it is the fifth version of my website (index_v5). I will do another upload and pull request with some stakeholder feedback for this stage. That will make the sixth version of my website, which will be named in Projectspace as index_v6.html and style_v6.css.