Upload Stage 3 v2 (index_v6)
This is the upload of the changes after getting stakeholder feedback at the end of Stage 3. It is Stage 3 v2; the second set of files for Images and Graphics. It is the sixth total version of the website (v6), and the Projectspace equivalent is index_v6.html, style_v6.css, and script_v3.js