Upload v7 (index_v7)
This is the upload of the code for the final completed version of the website, after the end of Stage 3. It adds responsive features and cleans up the last bits. It is the seventh total version of the website (index_v7), and the Projectspace equivalent is index_v7.html, style_v7.css, and script_v4.js